qgsRound: places should be of integer type instead of double

+++ src/core/qgis.h
@@ -301,7 +301,7 @@ inline bool qgsDoubleNearSig( double a, double b, int significantDigits = 10 )
  *
  * \since QGIS 3.0
  */
-inline double qgsRound( double number, double places )
+inline double qgsRound( double number, int places )
 {
   double m = ( number < 0.0 ) ? -1.0 : 1.0;
   double scaleFactor = std::pow( 10.0, places );
